Harden LCD summary source registry reads
